// client/src/modules/chatModel.ts
export interface ChatModel {
    
    clinicaId: string;

    callerSocketId: string;
    callerUserId: string;
    callerName: string;
    responderSocketId: string;
    responderUserId: string;
    responderName: string;
    callerDate: Date;
    responderDateDate: Date;
    callerMessage: string;
    responderrMessage: string;

}